+++ src/controllers/userInterestChannelController.js
@@ -1,0 +1,82 @@
+import {
+  findInterestChannelsById,
+  registerInterestChannel,
+  removeInterestChannel,
+  updateInterestChannelList,
+} from "../services/userInterestChannelService.js";
+
+export const getInterestChannelsById = async (req, res, next) => {
+  try {
+    const { userId } = req.params;
+    const channels = await findInterestChannelsById(Number(userId));
+
+    res.status(200).json(channels);
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const createInterestChannel = async (req, res, next) => {
+  try {
+    const { userId } = req.params;
+    const { channelId } = req.body;
+
+    if (typeof channelId !== "number") {
+      return res.status(400).json({
+        status: 400,
+        error: "올바르지 않은 형식입니다.",
+      });
+    }
+
+    const message = await registerInterestChannel(
+      Number(userId),
+      Number(channelId)
+    );
+
+    res.status(200).json(message);
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const updateInterestChannels = async (req, res, next) => {
+  try {
+    const { userId } = req.params;
+    const { channelIds } = req.body;
+
+    if (!Array.isArray(channelIds)) {
+      return res.status(400).json({
+        status: 400,
+        error: "올바르지 않은 형식입니다.",
+      });
+    }
+
+    const result = await updateInterestChannelList(Number(userId), channelIds);
+    res.status(200).json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const deleteInterestChannel = async (req, res, next) => {
+  try {
+    const { userId } = req.params;
+    const { channelId } = req.body;
+
+    if (typeof channelId !== "number") {
+      return res.status(400).json({
+        status: 400,
+        error: "올바르지 않은 형식입니다.",
+      });
+    }
+
+    const message = await removeInterestChannel(
+      Number(userId),
+      Number(channelId)
+    );
+
+    res.status(200).json(message);
+  } catch (error) {
+    next(error);
+  }
+};
